+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [Unreleased]
 
+- Implemented qwerty keyboard (swipe space bar to switch)
 - Implemented horizontal swiping of a key